requirements/base.txt
django==4.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
requests==2.31.0
djangorestframework==3.14.0
django-cors-headers==3.13.0